Separate MyFileChooser from guiutils.h
