+++ rtgui/CMakeLists.txt
@@ -86,6 +86,7 @@ set(RT_GUI_WIDGET_SOURCES
     widgets/basic/lwbutton.cc
     widgets/basic/lwbuttonset.cc
     widgets/basic/myexpander.cc
+    widgets/basic/myfilechooser.cc
     widgets/basic/popupbutton.cc
     widgets/basic/popupcommon.cc
     widgets/basic/popuptogglebutton.cc
