+++ rtgui/tools/icmpanel.cc
@@ -27,9 +27,10 @@
 #include "options.h"
 #include "pathutils.h"
 #include "rtimage.h"
+#include "widgets/basic/myexpander.h"
+#include "widgets/basic/myfilechooser.h"
 #include "widgets/curves/curveeditor.h"
 #include "widgets/curves/curveeditorgroup.h"
-#include "widgets/basic/myexpander.h"
 
 #include "rtengine/dcp.h"
 #include "rtengine/iccstore.h"
