+++ rtgui/tools/lensprofile.cc
@@ -28,6 +28,8 @@
 #include "guiutils.h"
 #include "rtimage.h"
 #include "options.h"
+#include "widgets/basic/myexpander.h"
+#include "widgets/basic/myfilechooser.h"
 
 #include "rtengine/lcp.h"
 #include "rtengine/procparams.h"
